Add infrastructure for unit testing.
MPI means that most of Gadget is impossible to test. However,
there are some parts which are basically serial and can be unit tested.
This commit adds the makefile infrastructure for this, without actually
adding any tests.

The unit testing framework used is cmocka, because it is fast, C-only and
reasonably simple.

Tests for $module, with C source in ${module}.c should be added as:
test/${module}_test.c

and ${src} added to the TESTED variable in Makefile.rules.

Tests are run with 'make test'