Fix stopping condition in density.c
density.c stops automatically if the different between the upper and
lower HSML bounds < 0.1 % of the lower bound. However, this is too large
a difference: with a lot of particles and neighbours it may be hit
when the algorithm just hasn't converged yet. Make the convergence limit
10^-6 instead and warn if it happens. If it is happening it is almost
certainly because multiple particles are at the same position, in which
case you may have a different bug that you want to know about.