Correct the code that shrinks the ngb bounds
The previous code was using desnumgb - MaxNumNgbDeviation, which
could lead to the upper bound not being an upper bound. This was
previously hidden by the condition removed in e12e175 which stopped
ngb convergence when the upper and lower bounds got too close. This
would have just given the wrong answer, after e12e175 it fails to
converge.

The code in Gadget-2 looks a little different and I am unsure what it is
doing. I guess it could be correct? But this might also be broken in
Gadget-2.